pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
requests = "^2.31.0"
flask = "^3.0.1"
openai = "^1.10.0"
packaging = "^23.2"
flask-cors = "^4.0.1"
google-api-python-client = "^2.132.0"
gspread = "^6.1.2"
flask-limiter = "^3.7.0"
google-auth = "^2.30.0"
pytz = "^2024.1"
user-agents = "^2.2.0"
gevent = "^24.2.1"
aiohttp = "^3.9.5"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"